+++ src/core/computedvalue.ts
@@ -130,7 +130,7 @@ export class ComputedValue<T> implements IObservable, IComputedValue<T>, IDeriva
                             .name}' is being read outside a reactive context and doing a full recompute`
                     )
                 }
-                if (globalState.warnOnUnsafeComputationReads) {
+                if (globalState.computedRequiresReaction) {
                     console.warn(
                         `[mobx] Computed value ${this
                             .name} is read outside a reactive context and not actively observed. Doing a full recompute`